+++ packages/orq-rc/RELEASES.md
@@ -8719,3 +8719,13 @@ Based on:
 - [typescript v4.9.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.9 - packages/orq-rc
+
+## 2026-04-30 05:31:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.10 - packages/orq-rc